Lab7

+++ app/Http/Controllers/ArticleController.php
@@ -7,6 +7,7 @@ use App\Models\User;
 use App\Models\Comment;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Gate;
 
 
 class ArticleController extends Controller
@@ -33,6 +34,7 @@ class ArticleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', [self::class]);
         $request->validate([
             'date'=>'date',
             'name'=>'required|min:5|max:100',
@@ -71,17 +73,18 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
-        $request->validate([
-            'date'=>'date',
-            'name'=>'required|min:5|max:100',
-            'desc'=>'required|min:5'
-        ]);
-        $article->date = $request->date;
-        $article->name = $request->name;
-        $article->desc = $request->desc;
-        $article->user_id = 1;
-        if ($article->save()) return redirect('/article')->with('status','Update success');
-        else return redirect()->route('article.index')->with('status','Update don`t success');
+        Gate::authorize('update', $article->id);
+            $request->validate([
+                'date'=>'date',
+                'name'=>'required|min:5|max:100',
+                'desc'=>'required|min:5'
+            ]);
+            $article->date = $request->date;
+            $article->name = $request->name;
+            $article->desc = $request->desc;
+            $article->user_id = 1;
+            if ($article->save()) return redirect('/article')->with('status','Update success');
+            else return redirect()->route('article.index')->with('status','Update don`t success');
     }
 
     /**
@@ -89,6 +92,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('delete', [self::class]);
         if ($article->delete()) return redirect('/article')->with('status','Delete success');
         else return redirect()->route('article.show', ['article'=>$article->id])->with('status','Delete don`t success');
     }
